requirements.txt
tmdbv3api==1.9.0
python-dotenv==1.0.1
requests==2.32.3
